atualizando rotas bugadas

+++ routes/imagem.py
@@ -23,12 +23,20 @@ def upload_imagem():
     except Exception as e:
         return jsonify({"error": "Erro ao processar a imagem: " + str(e)}), 500
 
-@imagem_bp.route('/delete/<string:nome_arquivo>', methods=['DELETE'])
-def delete_imagem(nome_arquivo):
+from flask import request
+
+@imagem_bp.route('/delete', methods=['POST'])  # Mudado para POST
+def delete_imagem():
     try:
+        data = request.get_json()  # Obtém os dados do corpo da requisição
+        nome_arquivo = data.get('link')  # Supondo que o link da imagem seja passado com a chave 'link'
+        print(nome_arquivo)
+        if not nome_arquivo:
+            return jsonify({"error": "Link da imagem não fornecido"}), 400
+
         # Usa o método delete da biblioteca vercel_blob para remover a imagem
         vercel_blob.delete(nome_arquivo)  # Remove o arquivo do Blob
         
-        return jsonify({"message": "Imagem deletada com sucesso"}), 204
+        return jsonify({"message": "Imagem deletada com sucesso", "success": True}), 200
     except Exception as e:
         return jsonify({"error": "Erro ao deletar a imagem: " + str(e)}), 500
